app/requirements.txt
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
djoser==2.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
cryptography==2.8.0
django-cors-headers==3.7.0